Refactor battle engines to abstract class to contain generic logic
